+++ R/plot.hmm.R
@@ -400,17 +400,23 @@ plot.hmm  <- function(x, layout = "horizontal", pie = TRUE,
   if (length(cpal) == 1 && cpal == "auto") {
     pie.colors  <- attr(x$observations, "cpal")
   } else if(length(cpal) != ncol(x$emiss)) {
-    warning("The length of the vector provided for argument cpal does not match the number of observed states. Automatic color palette was used.")
-    pie.colors  <- attr(x$observations, "cpal")
+    if(ncol(x$emiss) <= 200) {
+      warning("The length of the vector provided for argument cpal does not match the number of observed states. Automatic color palette was used.")
+      pie.colors  <- attr(x$observations, "cpal")
+    } else {
+      stop("Not enough colours in the colour palette. Argument 'cpal' should be of length ", ncol(x$emiss), ".")
+    }
   } else if(!all(isColor(cpal))) {
-    stop(paste("Please provide a vector of colors for argument cpal or use value \"auto\" for automatic color palette."))
+    stop("Please provide a vector of colors for argument cpal or use value \"auto\" for automatic color palette.")
   } else {
     pie.colors  <- cpal
   }
   if (with.legend != FALSE) {
     pie.colors.l  <- pie.colors
   }
-
+  if (length(pie.colors) != ncol(x$emiss)) {
+    stop("Not enough colours in the default colour palette. Argument 'cpal' should be of length ", ncol(x$emiss), ".")
+  }
   # Legend position and number of columns
   if (with.legend != FALSE && pie == TRUE) {
     # Own labels in legend